Fix indentation, clean, and add minor tweaks
The `printf` command is similar to but far superior and more portable
than `echo`, so I encourage its use. In any case, while `date` is a
fine enough option here, `printf` is a more efficient choice, since
it's built into the shell (bash).

Your formatting needed adjusting. The more by-the-book approach to
format blocks of code you used, would be as follows:

  if condition
  then
  	commands
  fi

  name()
  {
  	commands
  }

While there can be variations, and this isn't Python (strict indenting)
it's still important not to get too wild, lest your code be a nightmare
to read, let alone to edit. Also, keep it consistent.

Another common approach to the above, and my personal preference:

  if condition; then
  	commands
  fi

  name(){
  	commands
  }