Add bars() for quicker & easier underlining
So instead of copy-pasting or typing the `echo` bit yourself, with the
right amount of equal-signs, you just type `bars` and it'll do it for
you. It also makes it so much easier if you decide to change the
characters used in your bars.

I re-indented the code within your initial function, but if this was
undesired, however ill-advised, then I apologise. Things like these
are good things to mention in your commit messages, by the way; if not
for me, maybe for others, and even yourself.

You are doing really well, by the way!